session_manager/requirements.txt
# Session Manager Dependencies
# VCA 1.0 - Phase 1

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0

# OpenAI API
openai==1.10.0

# Voice Activity Detection
webrtcvad==2.0.10

# Audio Processing
numpy==1.26.3
scipy==1.11.4

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging
colorlog==6.8.2

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1
